Metrics Diagnostics single line output
Instead of having multie line metrics:

25-7-2017 07:40:44 Metrics[
                          classloading.loadedClassesCount=7,794
                          classloading.totalLoadedClassesCount=7,794

A single line metric is used:

25-7-2017 07:40:44 Metrics[client.endpoint.totalRegistrations=0]
25-7-2017 07:40:44 Metrics[cluster.clock.clusterStartTime=1,500,957,590,070]

This makes it a lot easier to parse and to correlate the metrics with time.

The single line diagnosticslogwriter is dropped completely since its only
purpose was dealing with the complex multiline metrics.
